+++ database/migrations/2017_10_31_141246_create_emp__phones_table.php
@@ -16,7 +16,7 @@ class CreateEmpPhonesTable extends Migration
         DB::statement("
             CREATE TABLE emp__phones (
                 Employee_ID int,
-                Phone_No int,
+                Phone_No bigint,
                 PRIMARY KEY(Employee_ID, Phone_No),
                 Foreign Key(Employee_ID) References employees (Employee_ID)
             )
